==> R/old code/2-analysis.Rmd <==
---
title: "Understanding the influence of transportation infrastrucure on Urban Heat in Phoenix, Arizona"
author: "C.G. Hoehne"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: pdf_document
---

```{r setup, include=FALSE}

# default to knitr option of code printing
knitr::opts_chunk$set(echo = TRUE)

# load fonts for use figures
#extrafont::loadfonts(device = "win") 

# list of all dependant packages
list.of.packages <- c("data.table", 
                      "lubridate",
                      "weathermetrics",
                      "rgdal",
                      "rgeos",
                      "raster",
                      "tmap",
                      "maptools",
                      "xtable",
                      "knitr",
                      "tinytex",
                      "here")

# install missing packages
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)

# install TinyTex (for creating PDFs from rmarkdown) if it is not installed.
# NOTE YOU NEED TO RESTART R STUDIO AFTER INSTALL
#tinytex::install_tinytex(force = FALSE)

# load packages
invisible(lapply(list.of.packages, library, character.only = TRUE)) # invisible() just hides printing stuff in console

# store loaded packages and thier version in a data.frame
package.info <- data.frame("Package" = list.of.packages, "Version" = unlist(lapply(list.of.packages, function (x) as.character(packageVersion(x)))))

# load windows fonts and store as string
windowsFonts(Century=windowsFont("TT Century Gothic"))
windowsFonts(Times=windowsFont("TT Times New Roman"))
my.font <- "Century"
#my.font <- "Times"

```

## Data Sources

### Weather Data
- National Centers for Environmental Information (NCEI): https://gis.ncdc.noaa.gov/maps/ncei/cdo/hourly
- The Arizona Meteorological Network (AZMET): https://cals.arizona.edu/azmet/az-data.htm
- Maricopa County Flood Control District (MCFCD): https://www.maricopa.gov/3769/Weather-Sensor-Data
- MesoWest: https://mesowest.utah.edu/

### Parking Data
- Forthcoming publication by Hoehne et al, info here: http://transportationlca.org/phoenixparking/

### Road Network Data
- OpenStreetMap (© OpenStreetMap contributors)
- Arizona Department of Transportation (ADOT)

### Traffic Data
- Calibrated Maricopa Association of Governments (MAG) travel demand and parcel data via modifed version of SUMO (Simulation of Urban MObility) using OpenStreetMap street network data


```{r data import, include = FALSE}

## import weather, roadway, traffic & parking data
w.data <- readRDS(here("data/outputs/temp/2017-weather-data.rds")) # cleaned hourly/sub-hourly weather data for all stations
stations.buffered <- readRDS(here("data/outputs/temp/station-buffers-final.rds")) # buffered station data w/ roadway & parking areas
radii.buffers <- readRDS(here("data/outputs/temp/radii-buffers.rds"))

# import cleaned traffic data
traffic <- fread(here("data/icarus_osm_traffic.csv"))

## import other various shapefiles for plotting
city.labels <- shapefile(here("data/shapefiles/other/phx_metro_labels.shp")) # city labels (pre-cropped to UZA)
uza.border <- shapefile(here("data/shapefiles/boundaries/maricopa_county_uza.shp")) # UZA 
cnty.border <- shapefile(here("data/shapefiles/boundaries/maricopa_county.shp")) # county 
highways <- shapefile(here("data/shapefiles/other/phx_metro_hways.shp")) # 2017 highways (pre-cropped to UZA)

# get station data as list of data.tables
uza.stations <- lapply(1:length(stations.buffered), function(x) as.data.table(stations.buffered[[x]]@data))

# filter weather data to all unique uza stations
all.station.ids <- unique(lapply(1:length(stations.buffered), function(x) unique(stations.buffered[[x]]@data$id)))[[1]]
uza.weather <- w.data[id %in% all.station.ids] 
rm(w.data)
gc()

# check unique stations
length(unique(uza.weather$id))
mean(sapply(1:length(stations.buffered), function(x) length(unique(stations.buffered[[x]]$id))))
## ** 2 more stations in station.buffers than uza.stations????

# count significant digits (NOT VECTORIZED)
signif.dig <- function(x) length(gregexpr("[[:digit:]]", as.character(x))[[1]])

# count digits to left hand side of decimal
lhs.dig <- function(x) ifelse(abs(x) < 1, 0, nchar(as.character(as.integer(abs(x)))))

```


```{r analysis, include = FALSE}

## First Pass Linear Regression Analysis

# calc hourly mean
uza.weather[, all_month_hr_mean := mean(temp.f, na.rm = T), by = .(hour,month)]

# aggregate uza weather data to data.table of summary variables
t.agg <-
  uza.weather %>%
  group_by(id) %>%
  summarise(yr_max_temp = max(temp.f, na.rm = T),
            smr_min_temp = min(temp.f[month %in% c("Jun","Jul","Aug")], na.rm = T),
            smr_5pm_temp = mean(temp.f[hour == 17 & month %in% c("Jun","Jul","Aug")], na.rm = T),
            smr_7pm_temp = mean(temp.f[hour == 19 & month %in% c("Jun","Jul","Aug")], na.rm = T),
            smr_9pm_temp = mean(temp.f[hour == 21 & month %in% c("Jun","Jul","Aug")], na.rm = T),
            rel_inc_month_hr_mean = mean(temp.f - all_month_hr_mean, na.rm = T),
            smr_5pm_month_hr_mean = mean(temp.f[hour == 19 & month %in% c("Jun","Jul","Aug")] - 
                                           all_month_hr_mean[hour == 19 & month %in% c("Jun","Jul","Aug")])) %>%
  data.table() 

# force abt NaN, Inf, and -Inf, to NA
invisible(lapply(names(t.agg),function(.name) set(t.agg, which(is.infinite(t.agg[[.name]]) |
                                                               is.nan(t.agg[[.name]])), j = .name,value =NA)))

# bind summary data to uza.stations
for(y in 1:length(uza.stations)){
  uza.stations[[y]] <- merge(uza.stations[[y]], t.agg, by = "id")
}


# create empty lists to store linear regression model/data/plots, iterate through, summarize, and plot
model.in <- list() # list cleaned input data to models
models <- list() # list of model objects
models.dt <- list() # list of data.tables of output data from models
models.pl <- list() # list of model plots

# loop through each temperature summary variable and each buffer distance creating a regression between the 2 and output plots for each combo

# all variables and corresponding labels
variables <- colnames(t.agg)[2:length(t.agg)] # list of summary variables
variable.labs <- c("2017 Max Observed Temp (deg F)",
                   "2017 Min Summer Observed Temp (deg F)",
                   "2017 Mean 5pm Summer Temp (deg F)",
                   "2017 Mean 7pm Summer Temp (deg F)",
                   "2017 Mean 9pm Summer Temp (deg F)",
                   "2017 Mean Change in Month-hour Temps (deg F)",
                   "2017 Mean Change in 5pm Month-hour Temp (deg F)")

# remove stations with no pavement area detected
#uza.stations.adj <- lapply(1:length(uza.stations), function(x) uza.stations[[x]][pave.prct > 0] )

# filter to which you want to plot
variables <- variables[1:7]
variable.labs <- variable.labs[1:7]

for(variable.j in 1:length(variables)){ # each summary variable j
  for(buffer.i in 1:length(uza.stations.adj)){ # each buffer index i
  
    ## models
    # first check and remove outliers in y (3 sd outside)
    model.in[[buffer.i]] <- as.data.table(data.frame(y = unlist(uza.stations.adj[[buffer.i]][,mget(variables[variable.j])]),
                                                     x = uza.stations.adj[[buffer.i]]$pave.prct))
    mean.y <- mean(model.in[[buffer.i]]$y, na.rm = T)
    sd.y <- sd(model.in[[buffer.i]]$y, na.rm = T)
    model.in[[buffer.i]] <- model.in[[buffer.i]][(y > mean.y - sd.y*3) & (y < mean.y + sd.y*3),]
    
    models[[buffer.i]] <- lm(model.in[[buffer.i]]$y ~ model.in[[buffer.i]]$x)
    models.dt[[buffer.i]] <- as.data.table(data.frame(fit = signif(models[[buffer.i]]$fitted.values,6), 
                                                      x = signif(models[[buffer.i]]$model[,2],6), 
                                                      y = signif(models[[buffer.i]]$model[,1],6)))
    # confidence intervals for plot banding
    conf.i <- as.data.table(predict(models[[buffer.i]], interval="confidence", level = 0.95))
    conf.i[, `:=`(lwr = signif(lwr,6), upr = signif(upr,6), fit.f = as.factor(signif(fit,6)))][,fit := NULL]
    models.dt[[buffer.i]] <- merge(models.dt[[buffer.i]][,fit.f := as.factor(fit)], unique(conf.i), by = "fit.f", all.x = T)[,fit.f := NULL]
    
    ## plots
    # plot x/y max/mins
    range.x <- diff(range(uza.stations.adj[[buffer.i]]$pave.prct))
    range.y <- diff(range(unlist(uza.stations.adj[[buffer.i]][,mget(variables[variable.j])]), na.rm = T))
    min.x <- 0
    max.x <- signif(max(models.dt[[buffer.i]]$x) + (0.1 * range.x), 2)
    min.y <- floor(min(models.dt[[buffer.i]]$y) - (0.1 * range.y))
    max.y <- ceiling(max(models.dt[[buffer.i]]$y) + (0.1 * range.y))
    chng.x <- models.dt[[buffer.i]]$x[2] - models.dt[[buffer.i]]$x[1]
    chng.y <- models.dt[[buffer.i]]$fit[2] - models.dt[[buffer.i]]$fit[1]
    
    # model plots using ggplot
    models.pl[[buffer.i]] <- (ggplot()
          + geom_segment(aes(x = min.x, y = min.y, xend = max.x, yend = min.y))   # x border (x,y) (xend,yend)
          + geom_segment(aes(x = min.x, y = min.y, xend = min.x, yend = max.y))  # y border (x,y) (xend,yend)
          + geom_ribbon(aes(ymin = lwr, ymax = upr, x = x), fill = "grey30", alpha = 0.2, data = models.dt[[buffer.i]])
          + geom_line(aes(y = fit, x = x), data = models.dt[[buffer.i]], colour = "red")
          + geom_point(aes(y = y, x = x), data = models.dt[[buffer.i]])
          + scale_x_continuous(expand = c(0,0), limits = c(min.x,max.x))
          + scale_y_continuous(expand = c(0,0), limits = c(min.y,max.y))
          + labs(x = paste0("Fractional Surrounding Area of Road & Parking Pavement (r=",radii.buffers[[buffer.i]],"ft)"), y = variable.labs[variable.j])
          + geom_text(aes(x = 0.80*(max.x - min.x), y = min.y + 0.25*(max.y - min.y)), label = paste0("+10% pavement = ",if(chng.y/chng.x > 0){"+"},signif(0.1*(chng.y/chng.x),2),"F"), parse = F, family = my.font)
          + geom_text(aes(x = 0.85*(max.x - min.x), y = min.y + 0.18*(max.y - min.y)), label = paste("R^2 ==",signif(summary(models[[buffer.i]])$r.squared, 2)), parse = T, family = my.font)
          + geom_text(aes(x = 0.85*(max.x - min.x), y = min.y + 0.11*(max.y - min.y)), label = paste("p ==",signif(anova(models[[buffer.i]])$'Pr(>F)'[1], 2)), parse = T, family = my.font)
          + geom_text(aes(x = 0.85*(max.x - min.x), y = min.y + 0.04*(max.y - min.y)), label = paste("n ==",nrow(models.dt[[buffer.i]])), parse = T, family = my.font)
          + theme_minimal()
          + theme(text = element_text(family = my.font, size = 12),
                  plot.margin = margin(t = 10, r = 20, b = 10, l = 10, unit = "pt")))
    
    # save plots as png
    var.path <- here(paste0("figures/",variables[variable.j])) # string of new variable path sub-directory
    dir.create(var.path, showWarnings = F) # create variable sub-directory in "figures/" 
    ggsave(paste0("r",radii.buffers[[buffer.i]],"ft.png"), models.pl[[buffer.i]], 
           device = "png", path = var.path, scale = 1, width = 6.5, height = 5, dpi = 300, units = "in")
    
  }
}


```

## Plots

Here are the working plots:

```{r data plots, echo = FALSE}

#-#-#-#-#-#-#-#
# Data Plots #
#-#-#-#-#-#-#

# set knitr chunk options to put spaces between plots
opts_chunk$set(fig.ncol = 3, fig.showtext = T, fig.sep = c('\\newline','\\newline','\\newline')) # only workds for LaTex output

# store palette for plotting stations
my.palette <- c("#0C120C","#640D14","#C20114","#6D7275")

# plot all station points with UZA, highways, & city labels for context.
stations.plot <-   
  tm_shape(uza.border) + tm_borders(lwd = 0.8, lty = "solid", col = "grey40", alpha = 0.7) + tm_fill(col = "grey90") + # urbanized area (UZA) border & fill
  tm_shape(uza.stations, bbox = uza.stations) + tm_dots(col = "source", border.col = NULL, palette = my.palette, size = 0.3, alpha = 0.8, title = "Data Source") + # stations
  tm_shape(cnty.border) + tm_borders(lwd = 0.8, lty = "solid", col = "grey20", alpha = 0.7) + # county border
  tm_scale_bar(position = c(0.4,0.0), breaks = c(0,5,10,15,20), size = 0.90, color.light = "grey85") + # scalebar
  tm_compass(north = 0, type = "4star", size = 2, show.labels = 1, position = c(0.9,0.85)) + # compass
  tm_shape(highways) + tm_lines(lwd = 1.1, col = "grey40") + # highways
  tm_shape(city.labels) + tm_text("name", size = 0.7, fontfamily = my.font,  fontface = "bold.italic", just = "center") + # city labels
  tm_layout(fontfamily = my.font, fontface = "italic", bg.color = "grey95", title.size = 1.5, legend.title.size = 1.20, # theme & formatting
            legend.text.size = 0.7, title = "", title.position = c(0.2,0.94),
            legend.position = c(0.1,0.45), outer.margins = c(0,0,0,0), asp = 0)

tmap_save(stations.plot, filename = here::here("figures/stations.png")) # option to save
stations.plot # plot in report

##
# calculate range of temps and mean temp for urbanized stations
##

# if time is within 15 min of hour, round to hour, otherwise drop, then
# aggregate to the min, mean, max temp at each unique timestep for all stations
uza.weather <- uza.weather[abs(difftime(date.time, date.time.round, units = "min")) < 15]
uza.weather.agg <- uza.weather[, .(min.temp.f = min(temp.f, na.rm = T),
                             med.temp.f = median(temp.f, na.rm = T),
                             max.temp.f = max(temp.f, na.rm = T)), by = .(date.time.round)]

# hourly temp min/mean/max by 2017 month aggregate across all stations
uza.weather.month.agg <- uza.weather[, .(min.temp.f = min(temp.f, na.rm = T),
                                   med.temp.f = median(temp.f, na.rm = T),
                                   max.temp.f = max(temp.f, na.rm = T)), by = .(hour,month)]

# plot hourly temp.f range by month for 2017 for all stations
# Plot historical tempeature plots of metro regions
month.x.hour.p <- ggplot(uza.weather.month.agg, aes(x = hour, y = med.temp.f , group = month)) +
  geom_ribbon(aes(ymin = min.temp.f, ymax = max.temp.f),  fill = "grey50", alpha = 0.5) +
  geom_line(size = 1.2) +
  geom_segment(aes(x = 0, y = 25, xend = 23, yend = 25)) +
  geom_segment(aes(x = 0, y = 25, xend = 0, yend = 135)) +
  facet_wrap(~ month , scales = "fixed") +
  labs(title = "Hourly Temperature Ranges by month for Phoenix UZA, 2017", x = "Hour", y = "Temperature (deg F)") +
  scale_x_continuous(limits = c(0,23), breaks = c(0,6,12,18), expand = c(0,0)) +
  scale_y_continuous(limits = c(25,135), breaks = c(25,50,75,100,125), expand = c(0,0)) +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold", family = my.font, size = 12, colour = "black", hjust = 1),
        strip.text = element_text(family = my.font, size = 11, colour = "black"),
        axis.title = element_text(face = "bold", family = my.font, size = 12, colour = "black"),
        axis.text = element_text(family = my.font, size = 10, colour = "black"),
        axis.text.x = element_text(hjust = 0.25) # for some reason the text isn't defaulted to centered on the ticks
  )

month.x.hour.p
ggsave("phx-uza-hrly-by-month-temp.png", month.x.hour.p, device = "png", path = here("figures"), scale = 1, width = 6.5, height = 8, dpi = 300, units = "in")

# plot hourly temp.f range for summer months for 2017 for all stations
uza.weather.sum.month.agg <- uza.weather[, .(min.temp.f = min(temp.f, na.rm = T),
                                   med.temp.f = median(temp.f, na.rm = T),
                                   max.temp.f = max(temp.f, na.rm = T)), by = .(hour,month,source)]
uza.weather.sum.month.agg <- uza.weather.sum.month.agg[month %in% c("Jun","Jul","Aug")]

sum.month.x.hour.p <- ggplot(uza.weather.sum.month.agg, aes(x = hour, y = med.temp.f , group = month)) +
  geom_ribbon(aes(ymin = min.temp.f, ymax = max.temp.f),  fill = "grey50", alpha = 0.5) +
  geom_line(size = 1.2) +
  geom_segment(aes(x = 0, y = 25, xend = 23, yend = 25)) +
  geom_segment(aes(x = 0, y = 25, xend = 0, yend = 135)) +
  facet_wrap(month ~ source, scales = "fixed") +
  labs(title = "Hourly Summer Temperatures Ranges for Phoenix UZA by Data Source, 2017", x = "Hour", y = "Temperature (deg F)") +
  scale_x_continuous(limits = c(0,23), breaks = c(0,6,12,18), expand = c(0,0)) +
  scale_y_continuous(limits = c(25,135), breaks = c(25,50,75,100,125), expand = c(0,0)) +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold", family = my.font, size = 12, colour = "black", hjust = 1),
        strip.text = element_text(family = my.font, size = 11, colour = "black"),
        axis.title = element_text(face = "bold", family = my.font, size = 12, colour = "black"),
        axis.text = element_text(family = my.font, size = 10, colour = "black"),
        axis.text.x = element_text(hjust = 0.25) # for some reason the text isn't defaulted to centered on the ticks
  
  )

sum.month.x.hour.p
ggsave("phx-uza-hrly-summer-temp.png", sum.month.x.hour.p, device = "png", path = here("figures"), scale = 1, width = 6.5, height = 8, dpi = 300, units = "in")

# create summer data aggreate by station for GIF
uza.weather.summer.station.agg <- uza.weather[month %in% c("Jun","Jul","Aug"), .(min.temp.f = min(temp.f, na.rm = T),
                                   mean.temp.f = mean(temp.f, na.rm = T),
                                   max.temp.f = max(temp.f, na.rm = T)), by = .(hour,station.name)]

```


```{r analysis-proto, eval=FALSE, include=FALSE}

## analysis unfinished parts

# calculate obs.time.days such that it is the sum of 1/2 the time btwn the last obs and 1/2 the time btwn the next obs (coverage time)
for(i in 1:nrow(uza.weather)){
  
  # boundary conditions
  if(i == 1){
    t.lag <- uza.weather[i+1, date.time]
    uza.weather[i, obs.time.days := abs(difftime(date.time, t.lag, units = "days"))]}
  else if (i == nrow(uza.weather)){
    t.lead <- uza.weather[i-1, date.time]
    uza.weather[i, obs.time.days := abs(difftime(date.time, t.lag, units = "days"))]}
  
  # non-boundary condition
  else {
    t.lead <- uza.weather[i-1, date.time]
    t.lag <- uza.weather[i+1, date.time]
    uza.weather[i, obs.time.days := abs(difftime(t.lead, t.lag, units = "days"))]}
  
  # fix boundary issues when transitioning btwn stations
}

# use t.obs to calucate degree-days 

```


This document was compiled using R `r getRversion()`, R Markdown, and the following packages:  

```{r appendix, echo = FALSE, results = 'asis'}

# hide xtable global comment option
options(xtable.comment = FALSE)

# output table of packages used with packages listed in alphabetical order
print(xtable(package.info[order(package.info$Package),], row.names = NULL), include.rownames=FALSE)

```
